backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib